refactor(persistence): move entity classes from Internal/ to Entities/ folder
Move entity classes to a dedicated Entities/ subfolder across AI, Features,
and Templating EFCore modules to improve discoverability and separation from
internal service code.

Affected modules:
- Granit.AI.EntityFrameworkCore: AIUsageRecordEntity, AIWorkspaceEntity
- Granit.Features.EntityFrameworkCore: TenantFeatureOverride
- Granit.Templating.EntityFrameworkCore: TemplateCategoryEntity, TemplateRevisionEntity

Also fix stale test assertions from PR #717 (GranitPasskeyOptions.SectionName,
OpenIddict permission count).
