+++ src/Granit.AI.EntityFrameworkCore/Extensions/AIModelBuilderExtensions.cs
@@ -1,3 +1,4 @@
+using Granit.AI.EntityFrameworkCore.Entities;
 using Granit.AI.EntityFrameworkCore.Internal;
 using Microsoft.EntityFrameworkCore;
