+++ src/Granit.AI.EntityFrameworkCore/Internal/AIUsageRecordEntityConfiguration.cs
@@ -1,3 +1,4 @@
+using Granit.AI.EntityFrameworkCore.Entities;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
