+++ src/Granit.AI.EntityFrameworkCore/Internal/EfAIUsageQueryableProvider.cs
@@ -1,3 +1,4 @@
+using Granit.AI.EntityFrameworkCore.Entities;
 using Microsoft.EntityFrameworkCore;
 
 namespace Granit.AI.EntityFrameworkCore.Internal;
