+++ src/Granit.AI.EntityFrameworkCore/Internal/EfAIUsageStore.cs
@@ -1,4 +1,5 @@
 using Granit.AI.Diagnostics;
+using Granit.AI.EntityFrameworkCore.Entities;
 using Microsoft.EntityFrameworkCore;
 
 namespace Granit.AI.EntityFrameworkCore.Internal;
