+++ src/Granit.Features.EntityFrameworkCore/Internal/FeaturesDbContext.cs
@@ -1,4 +1,5 @@
 using Granit.DataFiltering;
+using Granit.Features.EntityFrameworkCore.Entities;
 using Granit.Features.EntityFrameworkCore.Extensions;
 using Granit.MultiTenancy;
 using Granit.Persistence.Extensions;
