+++ src/Granit.Features.EntityFrameworkCore/Internal/TenantFeatureOverrideConfiguration.cs
@@ -1,3 +1,4 @@
+using Granit.Features.EntityFrameworkCore.Entities;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
