+++ src/Granit.Templating.EntityFrameworkCore/Extensions/TemplatingModelBuilderExtensions.cs
@@ -1,3 +1,4 @@
+using Granit.Templating.EntityFrameworkCore.Entities;
 using Granit.Templating.EntityFrameworkCore.Internal;
 using Microsoft.EntityFrameworkCore;
