+++ src/Granit.Templating.EntityFrameworkCore/Internal/TemplateCategoryEntityConfiguration.cs
@@ -1,3 +1,4 @@
+using Granit.Templating.EntityFrameworkCore.Entities;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Metadata.Builders;
