+++ tests/Granit.OpenIddict.Tests/Options/OptionsTests.cs
@@ -45,7 +45,7 @@ public sealed class OptionsTests
         options.ServerDomain.ShouldBeEmpty();
         options.AuthenticatorTimeout.ShouldBe(TimeSpan.FromMinutes(5));
         options.ChallengeSize.ShouldBe(32);
-        GranitPasskeyOptions.SectionName.ShouldBe("OpenIddict:Passkeys");
+        GranitPasskeyOptions.SectionName.ShouldBe("Identity:Passkeys");
     }
 
     [Fact]
