+++ tests/Granit.Templating.EntityFrameworkCore.Tests/EfTemplateCategoryStoreTests.cs
@@ -1,5 +1,6 @@
 using Granit.Exceptions;
 using Granit.Guids;
+using Granit.Templating.EntityFrameworkCore.Entities;
 using Granit.Templating.EntityFrameworkCore.Internal;
 using Granit.Templating.Store;
 using Granit.Timing;
