+++ tests/Granit.Templating.EntityFrameworkCore.Tests/StoreTemplateResolverTests.cs
@@ -1,3 +1,4 @@
+using Granit.Templating.EntityFrameworkCore.Entities;
 using Granit.Templating.EntityFrameworkCore.Internal;
 using Granit.Templating.Keys;
 using Granit.Templating.Pipeline;
