+++ tests/Granit.Templating.EntityFrameworkCore.Tests/TemplateCacheEntryTests.cs
@@ -1,3 +1,4 @@
+using Granit.Templating.EntityFrameworkCore.Entities;
 using Granit.Templating.EntityFrameworkCore.Internal;
 using Granit.Templating.Pipeline;
 using Shouldly;
